feat: create a 'ship' class
The ship includes properties such as 'length',
'hits' (the number of times it was hit) and
'sunk'.

There is a 'hit()' method that increases the number
of 'hits'.

There is an 'isSunk()' method that checks whether
a ship is considered sunk.

Add a test to check for an arbitrary length and
the default values of 'hits' and 'sunk'.