feat: create a 'gameboard' class
The gameboard has properties such as coordinates 'x'
and 'y', and a 2D array.

The method 'getBoard()' populates the 2D array with
the correct coordinates.

The methods 'isMissing()', 'isShip()' and
'isDamagedShip()' check whether specific events
occurred at the given coordinates.

Add tests to check whether the array is being correctly
populated and whether the methods are working as intended.