feat: Create a folder for DOM components
The first module 'startGame' contains a function
that integrate some of the core elements of the
game and the user interface.

Everything from 'startGame' is related to the
start of the game.

Style some of the HTML and DOM elements related
to the gameboards.

Add a 'boardDom' method to 'gameboard' class.