feat: Add a 'strategia' function
The function is tied to the user interface. When
the user clicks a square on the gameboard, it returns
the coordinates.

Fix 'startGame' by removing the event listener for
the 'start' button. The 'startGame' function now
focuses solely on creating objects like the gameboard.

Fix the start button by giving it a dedicated file
with its own event listeners.

Improve the gameboard by adding a hover effect.