feat: Add a new function 'getShips'
The function is able to place a different ship
depending on the turn. For example, the first turn
the player is able to place a 'carrier', the next
turn it's a 'battleship' and so on.

Nothing new has been introduced in the 'ships' class,
the function is related to the user interface.